Canvas rendering
Interpolation is not used to render the image. It reduces the image
quality.
A buffer or secondary canvas (hidden) is used to write the image. The
canvas is then scaled and written into the visible canvas.

Change of css for demo

Demo shows the correct panel orientation

Translate the image using a transformation

The image is translated using a transformation instead of calculating
the origin.
This will be convenient in the future if it is decided to use the
actual origin of the image.

Possibility to invert the x axis

the x axis can be inverted to match the view from Slicer.
The orientation of the slice can be set.